Move chebyshev.h to libhelfem
