+++ libhelfem/CMakeLists.txt
@@ -6,8 +6,10 @@ configure_file("include/helfem.source.hh" "include/helfem")
 add_library(helfem STATIC
     src/helfem.cc
     src/utils.cc
+    src/chebyshev.cc
 )
 target_include_directories(helfem PUBLIC "${CMAKE_CURRENT_BINARY_DIR}/include/" "${ARMADILLO_INCLUDE_DIRS}")
+target_include_directories(helfem PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/src/")
 target_link_libraries(helfem PUBLIC ${ARMADILLO_LIBRARIES})
 
 message("include: ${CMAKE_CURRENT_BINARY_DIR}/include/")
