+++ src/atomic/inttest.cpp
@@ -16,7 +16,7 @@
 #include "quadrature.h"
 #include "../general/polynomial_basis.h"
 #include "../general/polynomial.h"
-#include "../general/chebyshev.h"
+#include "chebyshev.h"
 
 using namespace helfem;
