+++ src/diatomic/basis.cpp
@@ -17,7 +17,7 @@
 #include "quadrature.h"
 #include "../general/polynomial.h"
 #include "../general/polynomial_basis.h"
-#include "../general/chebyshev.h"
+#include "chebyshev.h"
 #include "../general/spherical_harmonics.h"
 #include "../general/gaunt.h"
 #include "../general/gsz.h"
