+++ src/diatomic/twodquadrature.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "twodquadrature.h"
-#include "../general/chebyshev.h"
+#include "chebyshev.h"
 #include "../general/lcao.h"
 #include "../general/model_potential.h"
 #include "../general/utils.h"
