+++ src/harmonic/quadrature.cpp
@@ -14,7 +14,7 @@
  * of the License, or (at your option) any later version.
  */
 #include "quadrature.h"
-#include "../general/chebyshev.h"
+#include "chebyshev.h"
 
 namespace helfem {
   namespace quadrature {
