+++ src/sadatom/basis.cpp
@@ -15,7 +15,7 @@
  */
 #include "basis.h"
 #include "../general/polynomial.h"
-#include "../general/chebyshev.h"
+#include "chebyshev.h"
 #include "../general/spherical_harmonics.h"
 #include "../general/gaunt.h"
 #include "../general/gsz.h"
